+++ soh/soh/resource/type/Skeleton.cpp
@@ -49,7 +49,7 @@ void SkeletonPatcher::RegisterSkeleton(std::string& path, SkelAnime* skelAnime) 
 void SkeletonPatcher::UnregisterSkeleton(SkelAnime* skelAnime) {
 
     // TODO: Should probably just use a dictionary here...
-    for (int i = 0; i < skeletons.size(); i++) 
+    for (size_t i = 0; i < skeletons.size(); i++) 
     {
         auto skel = skeletons[i];
